+++ resources/css/map-component.css
@@ -1,0 +1,29 @@
+.you-are-here {
+    position: relative;
+    width: 10px;
+    height: 10px;
+    background: #4285F4;
+    border-radius: 50%;
+}
+
+.you-are-here::after {
+    content: '';
+    position: absolute;
+    inset: 0;
+    border-radius: 50%;
+    background: rgba(66, 133, 244, 0.5);
+    animation: pulse-ring 2s infinite;
+}
+
+@keyframes pulse-ring {
+    0% { transform: scale(1); opacity: 0.7; }
+    100% { transform: scale(3); opacity: 0; }
+}
+
+.leaflet-control-custom:hover {
+    background-color: #f4f4f4 !important;
+}
+
+.leaflet-bottom.leaflet-right {
+    margin-bottom: 10px;
+}
